pyproject.toml
[tool.poetry]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.11"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.39"}
asyncpg = "^0.30.0"
alembic = "^1.15.1"
flake8 = "^7.1.2"
flake8-pyproject = "^1.2.3"
ruff = "^0.11.2"
black = "^25.1.0"
pytest-asyncio = "^0.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    ".idea",
    "__init__.py"
]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = [".", "src"]
